feat: replace triage modal overlay with inline page + rail button
Triage now navigates to /triage (sidebar visible) instead of blocking
the page with RitualOverlay. Auto-gate shows TriageReminderBanner as a
toast nudge; a persistent "Morning triage" button with a blue dot
appears in the sidebar rail when triage is pending. Extracted TriageFlow
from TriageModal so the logic is reusable without the overlay.
